+++ src/main/java/com/comeon/backend/meeting/command/infra/MeetingRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.comeon.backend.meeting.infrastructure;
+package com.comeon.backend.meeting.command.infra;
 
 import com.comeon.backend.meeting.command.domain.*;
 import com.querydsl.jpa.impl.JPAQueryFactory;
