+++ src/main/java/com/comeon/backend/meeting/presentation/MeetingController.java
@@ -1,16 +1,16 @@
-package com.comeon.backend.meeting.presentation.api;
+package com.comeon.backend.meeting.presentation;
 
 import com.comeon.backend.common.api.SliceResponse;
 import com.comeon.backend.common.security.JwtPrincipal;
-import com.comeon.backend.meeting.command.application.MeetingCommandService;
-import com.comeon.backend.meeting.command.application.MeetingMemberCommandService;
-import com.comeon.backend.meeting.command.application.MeetingMemberSummary;
-import com.comeon.backend.meeting.presentation.api.request.MeetingAddRequest;
-import com.comeon.backend.meeting.presentation.api.request.MeetingJoinRequest;
-import com.comeon.backend.meeting.presentation.api.request.MeetingSummaryListParam;
-import com.comeon.backend.meeting.presentation.api.response.MeetingAddResponse;
-import com.comeon.backend.meeting.presentation.api.response.MeetingJoinResponse;
-import com.comeon.backend.meeting.presentation.api.response.MeetingSummaryResponse;
+import com.comeon.backend.meeting.command.application.MeetingFacade;
+import com.comeon.backend.meeting.command.application.MeetingMemberFacade;
+import com.comeon.backend.meeting.command.application.dto.MeetingMemberSummary;
+import com.comeon.backend.meeting.presentation.request.MeetingAddRequest;
+import com.comeon.backend.meeting.presentation.request.MeetingJoinRequest;
+import com.comeon.backend.meeting.presentation.request.MeetingSummaryListParam;
+import com.comeon.backend.meeting.presentation.response.MeetingAddResponse;
+import com.comeon.backend.meeting.presentation.response.MeetingJoinResponse;
+import com.comeon.backend.meeting.presentation.response.MeetingSummaryResponse;
 import com.comeon.backend.meeting.query.dao.MeetingDao;
 import com.comeon.backend.meeting.query.dao.MeetingCondition;
 import com.comeon.backend.meeting.query.dao.result.FindMeetingSliceResult;
@@ -32,14 +32,14 @@ import java.time.format.DateTimeFormatter;
 @RequestMapping("/api/v1/meetings")
 public class MeetingController {
 
-    private final MeetingCommandService meetingCommandService;
-    private final MeetingMemberCommandService meetingMemberCommandService;
+    private final MeetingFacade meetingFacade;
+    private final MeetingMemberFacade meetingMemberFacade;
     private final MeetingDao meetingDao;
 
     @PostMapping
     public MeetingAddResponse meetingAdd(@AuthenticationPrincipal JwtPrincipal jwtPrincipal,
                                          @Validated @RequestBody MeetingAddRequest request) {
-        Long meetingId = meetingCommandService.addMeeting(
+        Long meetingId = meetingFacade.addMeeting(
                 jwtPrincipal.getUserId(),
                 request.getMeetingName(),
                 request.getMeetingImageUrl(),
@@ -90,7 +90,7 @@ public class MeetingController {
     @PostMapping("/join")
     public MeetingJoinResponse meetingJoin(@AuthenticationPrincipal JwtPrincipal jwtPrincipal,
                                            @Validated @RequestBody MeetingJoinRequest request) {
-        MeetingMemberSummary meetingMemberSummary = meetingMemberCommandService.joinMeeting(jwtPrincipal.getUserId(), request.getEntryCode());
+        MeetingMemberSummary meetingMemberSummary = meetingMemberFacade.joinMeeting(jwtPrincipal.getUserId(), request.getEntryCode());
         return new MeetingJoinResponse(meetingMemberSummary.getMeetingId(), meetingMemberSummary.getMemberId(), meetingMemberSummary.getMemberRole());
     }
 }
